%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/mo.R
\name{mog_alcohol_licenses}
\alias{mog_alcohol_licenses}
\title{Missouri active alcohol licenses}
\usage{
mog_alcohol_licenses(city = NULL, county = NULL,
  license_type = NULL, limit = 1000)
}
\value{
tibble of alcohol license records
}
\description{
Missouri active alcohol licenses
}

