%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/mo.R
\name{mog_employee_pay}
\alias{mog_employee_pay}
\title{Missouri state employee pay}
\usage{
mog_employee_pay(year = 2025, agency = NULL, limit = 1000)
}
\value{
tibble of employee pay records
}
\description{
Missouri state employee pay
}

