%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/mo.R
\name{mog_expenditures}
\alias{mog_expenditures}
\title{Missouri state expenditures}
\usage{
mog_expenditures(year = 2025, limit = 1000)
}
\value{
tibble of expenditure records
}
\description{
Missouri state expenditures
}

