%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/mo.R
\name{mog_list}
\alias{mog_list}
\title{List Missouri Open Data datasets}
\usage{
mog_list(limit = 100)
}
\value{
tibble: id, name, category, description, updated_at
}
\description{
List Missouri Open Data datasets
}

