%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/mo.R
\name{mog_unemployment_claims}
\alias{mog_unemployment_claims}
\title{Missouri weekly initial unemployment claims}
\usage{
mog_unemployment_claims(county = NULL, start_date = NULL,
  end_date = NULL, limit = 1000)
}
\value{
tibble: weekending, county, claims
}
\description{
Missouri weekly initial unemployment claims
}

