%% Generated by roxygen2: do not edit by hand
%% Please edit documentation in R/mo.R
\name{mog_view}
\alias{mog_view}
\title{Query any Missouri dataset via SODA}
\usage{
mog_view(view_id, where = NULL, select = NULL, group = NULL,
  order = NULL, q = NULL, limit = 1000, offset = 0)
}
\value{
tibble of query results
}
\description{
Query any Missouri dataset via SODA
}

